refactor: moar into common dir
