+++ src/app/common/content-page/content-page.component.ts
