+++ src/app/common/tabs/tabs.component.scss
