+++ src/app/common/tabs/tabs.component.ts
@@ -12,7 +12,7 @@ import {
   viewChild,
   WritableSignal,
 } from '@angular/core'
-import { ToolbarButtonComponent } from '../toolbar-button/toolbar-button.component'
+import { ToolbarButtonComponent } from '@/common/toolbar-button/toolbar-button.component'
 import {
   KeyboardDoubleArrowLeft,
   KeyboardDoubleArrowRight,
