+++ src/app/common/toolbar-button/toolbar-button.component.scss
@@ -1,7 +1,7 @@
-@use 'toolbar';
-@use 'touch-or-pointer';
-@use 'material-symbols';
-@use 'animations';
+@use '../../../sass/toolbar';
+@use '../../../sass/touch-or-pointer';
+@use '../../../sass/material-symbols';
+@use '../../../sass/animations';
 
 :host {
   display: inline flex;
