+++ src/app/common/toolbar-divider/toolbar-divider.component.scss
@@ -1,5 +1,5 @@
-@use 'animations';
-@use 'margins';
+@use '../../../sass/animations';
+@use '../../../sass/margins';
 
 :host {
   display: inline;
