+++ src/app/resume-page/plain-resume/plain-resume.component.ts
@@ -3,7 +3,7 @@ import { LinkComponent } from '../link/link.component'
 import { DateRangeComponent } from '../date-range/date-range.component'
 import { EDUCATION_SERVICE } from '../data/education-service'
 import { PROJECT_SERVICE } from '../data/project-service'
-import { ContentPageComponent } from '../../content-page/content-page.component'
+import { ContentPageComponent } from '@/common/content-page/content-page.component'
 import { MdLinksPipe } from '../md-links.pipe'
 import { MaterialSymbolDirective } from '@/common/material-symbol.directive'
 import { NgIcon } from '@ng-icons/core'
